backend/requirements.txt
fastapi[standard]>=0.115.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
python-multipart>=0.0.18
instagrapi>=2.1.0
pillow>=11.0.0
apscheduler>=3.10.0
pydantic[email]>=2.0.0
google-genai>=0.3.0